// src/lib/services/compiler/types.ts
export interface CompileResult {
	js: string;
	css?: string;
	warnings: CompileWarning[];
	error?: CompileError;
}

export interface CompileWarning {
	code: string;
	message: string;
	filename?: string;
	pos?: number;
	start?: { line: number; column: number };
	end?: { line: number; column: number };
}

export interface CompileError extends Error {
	code?: string;
	filename?: string;
	pos?: number;
	start?: { line: number; column: number };
	end?: { line: number; column: number };
}

export interface CompilerOptions {
	filename?: string;
	css?: 'injected' | 'external' | 'none';
	generate?: 'dom' | 'ssr' | false;
	hydratable?: boolean;
	immutable?: boolean;
}